Move binaryfile-related tests to correct file
